+++ CHANGES.md
@@ -50,6 +50,7 @@ En vez de por área por responsable
     - Se crean los graficos para AMy NC totales y por origen
     - Creo filtros de año + fecha de apertura
     - Se pasan todas la generación de gráficos a jquery + ajax.post
+    - Aplicar filtro de todas las fechas de apertura para un año específico. (por si elije todas las fechas)
 
 
 ## FIX-ISO9K1